requirements.txt
notion-client
python-dotenv
yfinance